displayd/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
gpiod==2.1.3
idna==3.7
nats-py==2.7.2
netifaces==0.11.0
PyYAML==6.0.1
requests==2.31.0
sdnotify==0.3.2
smbus2==0.4.3
urllib3==2.2.1
